+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Improvements
+
+- [#961](https://github.com/cosmos/iavl/pull/961) Add new `GetLatestVersion` API to get the latest version.
+
 ## v1.2.0 May 13, 2024
 
 ### Improvements